fix(ci): make 'test' job pass on py3.12/3.13 — ruff fully clean + 67 tests green
- store.py: persistent :memory: conn in EventStore context (flush no longer
  hits 'no such table')
- events.py: fp field 15 uses version nibble (fingerprint = 16 bytes)
- config.example: storage paths on 3rd line
- CLI: load device from config.toml, never hard-coded
- CI: actions run on Node 24 (silences deprecation warnings)
- ruff: extend-exclude docs, BLE001/S110 per-file ignores for src,
  fix I001/RUF005/RUF012/RUF046/SIM103/SIM115-F841/UP031 in tree+tests
- ci.yml: explicit uses-node-version=24 for checkout/setup-python

+++ src/bruittrack/dsp.py
@@ -13,7 +13,6 @@ from __future__ import annotations
 
 import numpy as np
 
-
 try:  # optional fast path — see docs/decision-log.md
     from scipy.signal import sosfilt
 
@@ -326,7 +325,7 @@ def compute_channel_delay_ms(
     Positive delay: Left leads Right.
     Negative delay: Right leads Left.
     """
-    max_lag_samples = int(round(max_lag_ms * fs_low / 1000.0))
+    max_lag_samples = round(max_lag_ms * fs_low / 1000.0)
     if max_lag_samples < 1:
         return 0.0
